+++ src/datasources/data-frame/DataFrameDataSourceBase.ts
@@ -12,33 +12,35 @@ import {
     TableDataRows,
     Column
 } from './types';
-import { BackendSrv, getBackendSrv, TemplateSrv, getTemplateSrv } from '@grafana/runtime';
+import { BackendSrv, TemplateSrv } from '@grafana/runtime';
 
-export abstract class DataFrameDataSourceBase extends DataSourceBase<DataFrameQuery, DataFrameDataSourceOptions> {
-    baseUrl = this.instanceSettings.url + '/nidataframe/v1';
+export abstract class DataFrameDataSourceBase<
+    TQuery extends DataFrameQuery = DataFrameQuery,
+> extends DataSourceBase<TQuery, DataFrameDataSourceOptions> {
+    public baseUrl = this.instanceSettings.url + '/nidataframe/v1';
 
     public constructor(
-        readonly instanceSettings: DataSourceInstanceSettings<DataFrameDataSourceOptions>,
-        readonly backendSrv: BackendSrv = getBackendSrv(),
-        readonly templateSrv: TemplateSrv = getTemplateSrv()
+        public readonly instanceSettings: DataSourceInstanceSettings<DataFrameDataSourceOptions>,
+        public readonly backendSrv: BackendSrv,
+        public readonly templateSrv: TemplateSrv
     ) {
         super(instanceSettings, backendSrv, templateSrv);
     }
 
-    abstract processQuery(query: DataFrameQuery): ValidDataFrameQuery;
+    public abstract processQuery(query: TQuery): ValidDataFrameQuery;
 
-    abstract getTableProperties(id?: string): Promise<TableProperties>;
+    public abstract getTableProperties(id?: string): Promise<TableProperties>;
 
-    abstract getDecimatedTableData(
-        query: DataFrameQuery,
+    public abstract getDecimatedTableData(
+        query: TQuery,
         columns: Column[],
         timeRange: TimeRange,
         intervals?: number
     ): Promise<TableDataRows>;
 
-    abstract queryTables(query: string): Promise<TableProperties[]>;
+    public abstract queryTables(query: string): Promise<TableProperties[]>;
 
-    async testDatasource(): Promise<TestDataSourceResponse> {
+    public async testDatasource(): Promise<TestDataSourceResponse> {
         await this.get(`${this.baseUrl}/tables`, { take: 1 });
         return { status: 'success', message: 'Data source connected and authentication successful!' };
     }